Code reorganization for improved readability. Using an ephemeral url session now instead of default. Added provider option for setting customUserAgent. Added provider option for explicitly encoding requests into the http body.
